move matrix generation to its own file
