+++ packages/client/hmi-client/src/model-representation/petrinet/mira-petri.ts
@@ -1,7 +1,8 @@
 import _ from 'lodash';
 import { Model, PetriNetTransition } from '@/types/Types';
-import { createMatrix1D, createParameterMatrix, createTransitionMatrix } from '@/utils/pivot';
+import { createMatrix1D } from '@/utils/pivot';
 import { StratifiedMatrix } from '@/types/Model';
+import { createParameterMatrix, createTransitionMatrix } from './petri-matrix-service';
 
 /**
  * Note "id" and "base" used for building the compact graph, they should not be used as strata dimensions
